refactor!: (wip) split `crate::dispatch`
