+++ src/cmd.rs
@@ -1,19 +1,31 @@
 //! Definitions for command line argument mapping and dispatching.
+//!
+//! An overall introduction of how this module works:
+//! 1. [`clap`] handles command line arguments and generate a [`Pacaptr`]
+//!   instance holding all the flags and options.
+//! 2. [`Config`] reads the configuration file (if it exists) and then merges it
+//!   with the current command line arguments using [`Pacaptr::merge_cfg`].
+//! 3. The correct package manager to be used will be indicated by the user
+//!   (through command line arguments or config file), or, if this is not the
+//!   case, automatically detected by [`detect_pm_str`].
+//! 4. [`Pacaptr::dispatch`] will call the corresponding trait method, eg.
+//!   `.suy()`, according to the combination of flags and options obtained
+//!   above.
 
 use clap::{self, ArgAction, Parser};
 use figment::{providers::Serialized, Figment};
 use itertools::Itertools;
-use tap::prelude::*;
-use tokio::task;
-use tt_call::tt_call;
-
-use crate::{
-    _built::GIT_VERSION,
-    dispatch::Config,
+use pacaptr::{
+    config::Config,
     error::{Error, Result},
     methods,
     pm::BoxPm,
 };
+use tap::prelude::*;
+use tokio::task;
+use tt_call::tt_call;
+
+use crate::_built::GIT_VERSION;
 
 /// The command line options to be collected.
 #[derive(Debug, Parser)]
@@ -317,53 +329,10 @@ impl Pacaptr {
 
 #[cfg(test)]
 pub(super) mod tests {
-    use async_trait::async_trait;
     use once_cell::sync::Lazy;
     use tokio::test;
-    use tt_call::tt_call;
 
     use super::*;
-    use crate::pm::Pm;
-
-    pub struct MockPm {
-        pub cfg: Config,
-    }
-
-    macro_rules! make_mock_op_body {
-        ($self:ident, $kws:ident, $flags:ident, $method:ident) => {{
-            let kws: Vec<_> = itertools::chain!($kws, $flags).collect();
-            panic!("should run: {} {:?}", stringify!($method), &kws)
-        }};
-    }
-
-    macro_rules! impl_pm_mock {(
-        methods = [{ $(
-            $( #[$meta:meta] )*
-            async fn $method:ident;
-        )* }]
-    ) => {
-        #[async_trait]
-        impl Pm for MockPm {
-            /// Gets the name of the package manager.
-            fn name(&self) -> &str {
-                "mockpm"
-            }
-
-            fn cfg(&self) -> &Config {
-                &self.cfg
-            }
-
-            // * Automatically generated methods below... *
-            $( async fn $method(&self, kws: &[&str], flags: &[&str]) -> Result<()> {
-                    make_mock_op_body!(self, kws, flags, $method)
-            } )*
-        }
-    };}
-
-    tt_call! {
-        macro = [{ methods }]
-        ~~> impl_pm_mock
-    }
 
     static MOCK_CFG: Lazy<Config> = Lazy::new(|| Config {
         default_pm: Some("mockpm".into()),
