+++ src/lib.rs
@@ -20,12 +20,8 @@
     unused_allocation
 )]
 
-pub mod dispatch;
+pub mod config;
 pub mod error;
-mod exec;
-mod pm;
+pub mod exec;
+pub mod pm;
 pub mod print;
-
-mod _built {
-    include!(concat!(env!("OUT_DIR"), "/built.rs"));
-}
