+++ src/main.rs
@@ -1,5 +1,13 @@
+mod cmd;
+
+mod _built {
+    include!(concat!(env!("OUT_DIR"), "/built.rs"));
+}
+
 use clap::Parser;
-use pacaptr::{dispatch::Pacaptr, error::MainError};
+use pacaptr::error::MainError;
+
+use crate::cmd::Pacaptr;
 
 #[tokio::main]
 async fn main() -> Result<(), MainError> {
