+++ src/pm.rs
@@ -40,9 +40,9 @@ use paste::paste;
 use tt_call::tt_call;
 
 use crate::{
-    dispatch::Config,
+    config::Config,
     error::Result,
-    exec::{self, Cmd, Mode, Output},
+    exec::{self, is_exe, Cmd, Mode, Output},
     print::{println_quoted, prompt},
 };
 
@@ -232,11 +232,113 @@ pub trait Pm: Sync {
 /// An owned, dynamically typed [`Pm`].
 pub type BoxPm<'a> = Box<dyn Pm + Send + 'a>;
 
+impl From<Config> for BoxPm<'_> {
+    /// Generates the `Pm` instance according it's name, feeding it with the
+    /// current `Config`.
+    fn from(mut cfg: Config) -> Self {
+        // If the `Pm` to be used is not stated in any config,
+        // we should fall back to automatic detection and overwrite `cfg`.
+        let pm = cfg.default_pm.get_or_insert_with(|| detect_pm_str().into());
+
+        #[allow(clippy::match_single_binding)]
+        match pm.as_ref() {
+            // Chocolatey
+            "choco" => Choco::new(cfg).boxed(),
+
+            // Scoop
+            "scoop" => Scoop::new(cfg).boxed(),
+
+            // Winget
+            "winget" => Winget::new(cfg).boxed(),
+
+            // Homebrew/Linuxbrew
+            "brew" => Brew::new(cfg).boxed(),
+
+            // Macports
+            "port" if cfg!(target_os = "macos") => Port::new(cfg).boxed(),
+
+            // Apt for Debian/Ubuntu/Termux (newer versions)
+            "apt" => Apt::new(cfg).boxed(),
+
+            // Apk for Alpine
+            "apk" => Apk::new(cfg).boxed(),
+
+            // Dnf for RedHat
+            "dnf" => Dnf::new(cfg).boxed(),
+
+            // Portage for Gentoo
+            "emerge" => Emerge::new(cfg).boxed(),
+
+            // Xbps for Void Linux
+            "xbps" | "xbps-install" => Xbps::new(cfg).boxed(),
+
+            // Zypper for SUSE
+            "zypper" => Zypper::new(cfg).boxed(),
+
+            // -- External Package Managers --
+
+            // Conda
+            "conda" => Conda::new(cfg).boxed(),
+
+            // Pip
+            "pip" | "pip3" => Pip::new(cfg).boxed(),
+
+            // PackageKit
+            "pkcon" => Pkcon::new(cfg).boxed(),
+
+            // Tlmgr
+            "tlmgr" => Tlmgr::new(cfg).boxed(),
+
+            // Test-only mock package manager
+            #[cfg(test)]
+            "mockpm" => {
+                use self::tests::MockPm;
+                MockPm { cfg }.boxed()
+            }
+
+            // Unknown package manager X
+            x => Unknown::new(x).boxed(),
+        }
+    }
+}
+
+/// Detects the name of the package manager to be used in auto dispatch.
+#[must_use]
+fn detect_pm_str<'s>() -> &'s str {
+    let pairs: &[(&str, &str)] = match () {
+        _ if cfg!(target_os = "windows") => &[("scoop", ""), ("choco", ""), ("winget", "")],
+
+        _ if cfg!(target_os = "macos") => &[
+            ("brew", "/usr/local/bin/brew"),
+            ("port", "/opt/local/bin/port"),
+            ("apt", "/opt/procursus/bin/apt"),
+        ],
+
+        _ if cfg!(target_os = "ios") => &[("apt", "/usr/bin/apt")],
+
+        _ if cfg!(target_os = "linux") => &[
+            ("apk", "/sbin/apk"),
+            ("apt", "/usr/bin/apt"),
+            ("dnf", "/usr/bin/dnf"),
+            ("emerge", "/usr/bin/emerge"),
+            ("xbps-install", "/usr/bin/xbps-install"),
+            ("zypper", "/usr/bin/zypper"),
+        ],
+
+        _ => &[],
+    };
+
+    pairs
+        .iter()
+        .find_map(|&(name, path)| is_exe(name, path).then_some(name))
+        .unwrap_or("unknown")
+}
+
 /// Extra implementation helper functions for [`Pm`],
 /// focusing on the ability to run commands ([`Cmd`]s) in a configured and
 /// [`Pm`]-specific context.
 #[async_trait]
-trait PmHelper: Pm {
+pub trait PmHelper: Pm {
     /// Executes a command in the context of the [`Pm`] implementation. Returns
     /// the [`Output`] of this command.
     async fn check_output(&self, mut cmd: Cmd, mode: PmMode, strat: &Strategy) -> Result<Output> {
@@ -330,7 +432,7 @@ impl<P: Pm> PmHelper for P {}
 ///
 /// Default value: [`PmMode::CheckErr`].
 #[derive(Copy, Clone, Debug, Default)]
-enum PmMode {
+pub enum PmMode {
     /// Silently collects all the `stdout`/`stderr` combined. Print nothing.
     Mute,
 
@@ -360,7 +462,7 @@ impl From<PmMode> for Mode {
 /// package manager, indicating how it is run.
 #[derive(Clone, Debug, Default)]
 #[must_use]
-struct Strategy {
+pub struct Strategy {
     /// How a dry run is dealt with.
     dry_run: DryRunStrategy,
 
@@ -376,7 +478,7 @@ struct Strategy {
 /// Default value: [`DryRunStrategy::PrintCmd`].
 #[must_use]
 #[derive(Debug, Clone, Default)]
-enum DryRunStrategy {
+pub enum DryRunStrategy {
     /// Prints the command to be run, and stop.
     #[default]
     PrintCmd,
@@ -386,7 +488,7 @@ enum DryRunStrategy {
 
 impl DryRunStrategy {
     /// Invokes the corresponding package manager with the flags given.
-    fn with_flags(flags: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
+    pub fn with_flags(flags: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
         Self::WithFlags(flags.into_iter().map(|s| s.as_ref().into()).collect())
     }
 }
@@ -396,7 +498,7 @@ impl DryRunStrategy {
 /// Default value: [`PromptStrategy::None`].
 #[must_use]
 #[derive(Debug, Clone, Default)]
-enum PromptStrategy {
+pub enum PromptStrategy {
     /// There is no prompt.
     #[default]
     None,
@@ -413,13 +515,13 @@ enum PromptStrategy {
 impl PromptStrategy {
     /// There is a native prompt provided by the package manager
     /// that can be disabled with a flag.
-    fn native_no_confirm(no_confirm: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
+    pub fn native_no_confirm(no_confirm: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
         Self::NativeNoConfirm(no_confirm.into_iter().map(|s| s.as_ref().into()).collect())
     }
 
     /// There is a native prompt provided by the package manager
     /// that can be enabled with a flag.
-    fn native_confirm(confirm: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
+    pub fn native_confirm(confirm: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
         Self::NativeConfirm(confirm.into_iter().map(|s| s.as_ref().into()).collect())
     }
 }
@@ -429,7 +531,7 @@ impl PromptStrategy {
 /// Default value: [`PromptStrategy::None`].
 #[must_use]
 #[derive(Debug, Clone, Default)]
-enum NoCacheStrategy {
+pub enum NoCacheStrategy {
     /// Does not clean cache.
     /// This variant MUST be used when implementing cache cleaning methods like
     /// `-Sc`.
@@ -448,7 +550,57 @@ enum NoCacheStrategy {
 
 impl NoCacheStrategy {
     /// Invokes the corresponding package manager with the flags given.
-    fn with_flags(flags: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
+    pub fn with_flags(flags: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
         Self::WithFlags(flags.into_iter().map(|s| s.as_ref().into()).collect())
+    }
+}
+
+#[cfg(test)]
+pub mod tests {
+    use async_trait::async_trait;
+    use tt_call::tt_call;
+
+    use super::*;
+    use crate::config::Config;
+
+    #[derive(Debug)]
+    pub struct MockPm {
+        pub cfg: Config,
+    }
+
+    macro_rules! make_mock_op_body {
+        ($self:ident, $kws:ident, $flags:ident, $method:ident) => {{
+            let kws: Vec<_> = itertools::chain!($kws, $flags).collect();
+            panic!("should run: {} {:?}", stringify!($method), &kws)
+        }};
+    }
+
+    macro_rules! impl_pm_mock {(
+        methods = [{ $(
+            $( #[$meta:meta] )*
+            async fn $method:ident;
+        )* }]
+    ) => {
+        #[async_trait]
+        impl Pm for MockPm {
+            /// Gets the name of the package manager.
+            fn name(&self) -> &str {
+                "mockpm"
+            }
+
+            fn cfg(&self) -> &Config {
+                &self.cfg
+            }
+
+            // * Automatically generated methods below... *
+            $( async fn $method(&self, kws: &[&str], flags: &[&str]) -> Result<()> {
+                    make_mock_op_body!(self, kws, flags, $method)
+            } )*
+        }
+    };}
+
+    tt_call! {
+        macro = [{ methods }]
+        ~~> impl_pm_mock
     }
 }
